Cleaning headers
git-svn-id: https://csp.svn.sourceforge.net/svnroot/csp/trunk@45 7152ba15-5bd8-0310-b828-e896570a432b

+++ APPLICATIONS/CSPFlightSim/Include/BaseScreen.h
@@ -45,7 +45,7 @@ public:
     virtual void OnInit() = 0;
     virtual void OnExit() = 0;
 
-    virtual void OnUpdateObjects(float dt) = 0;
+    virtual void OnUpdateObjects(double dt) = 0;
     virtual void OnRender() = 0;
 
     // Input Routines, are not required. If the are not overloaded then
